# apiserver/model_device.go
/*
 * Loriot.io app API
 *
 * API to access and configure the Loriot.io app
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package apiserver

// Device - LoRaWAN device handled by the Loriot.io app
type Device struct {

	// App configs defining the Loriot.io target the device belongs to
	ConfigID *string `json:"configID,omitempty"`

	// Eliona project ID the asset belongs to
	ProjectID *string `json:"projectID,omitempty"`

	// Unique identifier for the asset
	GlobalAssetIdentifier string `json:"globalAssetIdentifier,omitempty"`

	// Application hexadecimal (uppercase) ID for Loriot
	AppId string `json:"appId,omitempty"`

	// Global ID in IEEE EUI64 address space that uniquely identifies the device
	DevEUI string `json:"devEUI,omitempty"`

	// corresponding asset ID
	AssetID *string `json:"assetID,omitempty"`
}

// AssertDeviceRequired checks if the required fields are not zero-ed
func AssertDeviceRequired(obj Device) error {
	return nil
}

// AssertDeviceConstraints checks if the values respects the defined constraints
func AssertDeviceConstraints(obj Device) error {
	return nil
}